+++ Ruddarr/Models/Instance.swift
@@ -17,39 +17,6 @@ enum InstanceType: String, Identifiable, CaseIterable, Codable {
 struct InstanceStatus: Codable {
   let appName: String
 }
-
-extension Array<Instance>: RawRepresentable {
-    public init?(rawValue: String) {
-        guard let data = rawValue.data(using: .utf8),
-            let result = try? JSONDecoder().decode([Instance].self, from: data)
-        else {
-            return nil
-        }
-        self = result
-    }
-
-    public var rawValue: String {
-        guard let data = try? JSONEncoder().encode(self),
-            let result = String(data: data, encoding: .utf8)
-        else {
-            return "[]"
-        }
-        return result
-    }
-}
-
-extension UUID: RawRepresentable {
-    public var rawValue: String {
-        self.uuidString
-    }
-
-    public typealias RawValue = String
-
-    public init?(rawValue: RawValue) {
-        self.init(uuidString: rawValue)
-    }
-}
-
 extension Instance {
     static var sample: Self {
         .init(
